chore: resolve the clang-tidy findings of the verification changes
The clang-tidy workflow treats every finding as an error. Five of them were on
lines the verification commits added:

- the tolerance of the projection tests cast a double to double in double
  precision builds; the float epsilon is promoted without the cast,
- readSampleTable filled its member list with push_back in a loop instead of
  sizing it up front,
- the run-files tests put their helpers into an anonymous namespace in a
  header; they are in a named one now,
- the deterministic point mapping reduced std::uint64_t with MPI_UINT64_T,
  which clang-tidy does not accept for unsigned long; it now asks
  Mpi::castToMpiType, as the rest of the code does,
- AsyncWriter used file::RunFiles without including its header.

AI-generated. Model: Opus 5

+++ tests/Numerical/Projection.t.h
@@ -35,7 +35,7 @@ constexpr double Tolerance = 1e-10;
 // What the code generator ships is stored in the precision of the build, so a comparison against it
 // cannot be tighter than that.
 constexpr double GeneratedTolerance =
-    std::max(Tolerance, 10.0 * static_cast<double>(std::numeric_limits<real>::epsilon()));
+    std::max(Tolerance, 10.0 * std::numeric_limits<real>::epsilon());
 
 // For fused simulations the code generator transposes everything in the `nodal` namespace
 // (cf. kernels/aderdg.py); detect that from a matrix whose shape is not square.
